Renamed quadratic/quadratic_results.py in quadratic/views.py, updated quadratic/urls.py
