+++ quadratic/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import patterns, url
 
-import quadratic_results
+import views
 
 urlpatterns = patterns('',
     # ex: /polls/
@@ -12,7 +12,7 @@ urlpatterns = patterns('',
     # ex: /polls/5/results/
     #url(r'^(?P<question_id>\d+)/results/$', views.results, name='results'),
 
-    url(r'^results/$', quadratic_results.results, name='results'),
+    url(r'^results/$', views.results, name='results'),
     # ex: /polls/5/vote/
     #url(r'^(?P<question_id>\d+)/vote/$', views.vote, name='vote'),
 )
